Move parameters to its own file
